// client/src/app/routes.ts
export enum AppRoutes {
  Splash = 'splash',
  Intro = 'intro',
  IntroAbout = 'intro/about',
  IntroTermsAndConditions = 'intro/terms',
  ChangeLanguage = 'languages/change',
  ListLanguages = 'languages',
  ViewLanguage = 'languages/:id',
  TermsAndConditions = 'terms',
  ImageSource = 'image/source',
  CaptureImage = 'image/capture',
  CaptionImage = 'image/caption',
  Translate = 'image/translate',
  Feedback = 'feedback',
  Technology = 'technology',
  AddWord = 'add-word',
  About = 'about',
  Unsupported = 'unsupported'
}